Restructure wordle blocks and grids
